Rename files to avoid conflicts during compilation
