+++ src/common/app_helper.h
