+++ src/common/tx.c
@@ -17,7 +17,7 @@
 #include "tx.h"
 #include "apdu_codes.h"
 #include "buffering.h"
-#include "parser.h"
+#include "parser_tx.h"
 #include <string.h>
 #include "zxmacros.h"
